+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changed
 - Renamed `PeerHandle::next_message()` to `recv_message()`.
 - Moved message body out of `ReceivedRequest`.
+- Renamed `Incoming` to `ReceivedMessage`.
 
 ## v0.4.2 - 2021-05-20
 ### Fixed